requirements.txt
annotated-types==0.7.0
asgiref==3.10.0
celery==5.4.0
cffi==2.0.0
contextlib2==21.6.0
cryptography==46.0.3
Django==5.2.7
django-cors-headers==4.9.0
django-ninja==1.5.0
django-ninja-extra==0.30.6
django-ninja-jwt==5.3.5
injector==0.23.0
openai==1.58.1
pillow==11.3.0
pydantic==2.12.1
pydantic_core==2.41.3
PyJWT==2.10.1
python-dotenv==1.2.1
redis==5.2.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
